requirements.txt
pefile>=2023.2.7
yara-python>=4.3.1
